pkgs/deploy/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "deploy"
version = "0.1.0"
dependencies = [
  "gitpython",
  "requests",
  "rich",
]

[project.scripts]
deploy = "main:main"
